requirements.txt
Django>=5.0,<6.0
python-dotenv>=1.0
dj-database-url>=2.2
gunicorn>=22.0
psycopg[binary]>=3.2
pytest>=8.0
pytest-django>=4.8
black>=24.0
ruff>=0.5
Pillow>=10.0
